Checkpoints tea 158 (#69)
* nearly there

* submodule and guid component

* guid cereal

* busted

* Forgot this lmao

* not breaking

* json not dardy

* big one

* idk why fsm not being found

* if rabbit deserialize can actually set the fsm it works

* so close

* checkpoints done fuck me

* update rabbit lua

* fixed switched to groups

---------

+++ Core/src/Controller/LuaAccess.cpp
@@ -419,7 +419,7 @@ namespace Vakol::Controller {
 
             auto& animation = ent->GetComponent<Components::Animation>();
 
-            animation.state = animation_state;
+            animation.state = static_cast<int>(animation_state);
         });
 
         entity_type.set_function("get_animation_duration", [](const Entity* ent, const int animation_state) {
